requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
cffi==1.16.0
Django==5.0.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-redis==5.4.0
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.4.1
psycopg2==2.9.9
pycparser==2.21
pytz==2024.1
redis==5.0.1
sqlparse==0.4.4
